src/scgrn/utils/seed.py
"""Seed helpers."""

from __future__ import annotations

import random

import numpy as np


def seed_everything(seed: int) -> None:
    random.seed(seed)
    np.random.seed(seed)
    try:
        import torch

        torch.manual_seed(seed)
        if torch.cuda.is_available():
            torch.cuda.manual_seed_all(seed)
        torch.set_float32_matmul_precision("medium")
    except Exception:
        pass

    try:
        import scvi

        scvi.settings.seed = seed
    except Exception:
        pass
