requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
anndata
scanpy
scvi-tools
pyscenic
loompy
pyyaml
